Rename FriendsCatalogTab, shouldHaveFriendRequestFrom
